[Refactor] change file name, (ssd_commands.write → ssd_commands.ssd_write, tests.test_write → tests.test_ssd_write)
